backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
numpy==1.26.4
supabase==2.13.0
google-generativeai==0.8.4
scikit-learn==1.6.1
sentence-transformers==3.4.1
pytest==8.3.4
httpx==0.28.1
